+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- [Minor] CPXE: Added more unittests
+- [Minor] CpxE and CpxAP: Added __repr__ function
 - [Minor] CpxE: Added property setter / getter for more modules.
 - [Minor] CpxE: Added access-by-name for modules of CpxE.
 - [Minor] CpxE: Added construction from typecode.
@@ -14,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Minor] CpxE8Do: Added access-by-index-operator to channels
 - [Minor] CpxE: Added CLI for writing and reading values.
 ### Changed
+- [Minor] changed return value of read function number to int
 - [Minor] several bugfixes
 - [Minor] Added context manager functionality to cpx base class. Adapted examples to use this context manager.
 - [Minor] Exchanged int_to_signed16 function with built in function
@@ -27,7 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Minor] Reworked CPX-E class structure.
 
 ### Fixed
-- [Patch] Bugfixes after next Merge
+- [Patch] Fixed diagnostics function in cpx-e-8do
+- [Patch] Fixed several linter issues
 - [Patch] Bugfixes after Merge
 
 ## v0.1.1 - 20.10.23